requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
geoip2==4.8.0
idna==3.10
laiarturs-ros-api==1.1.0
maxminddb==2.6.2
multidict==6.1.0
mysql-connector==2.2.9
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
yarl==1.17.1
